Lammort-Time wird jetzt in der Übung 3 verwendet

+++ AVA/AVALamportClock.swift
@@ -12,7 +12,7 @@ import Cocoa
 let LAMPORT_CLOCK = AVALamportClock.sharedInstance
 
 
-typealias AVALamportTimestamp = UInt
+typealias AVALamportTimestamp = UInt64
 
 
 class AVALamportClock: NSObject {
